+++ tekore/_auth/__init__.py
@@ -3,7 +3,10 @@ from .refreshing import RefreshingCredentials, RefreshingToken
 from .scope import scope, Scope
 from .token import Token, AccessToken
 from .util import (
+    UserAuth,
+    gen_state,
     parse_code_from_url,
+    parse_state_from_url,
     refresh_user_token,
     request_client_token,
     prompt_for_user_token,