+++ tekore/_auth/util.py
@@ -1,7 +1,34 @@
 import webbrowser
 
+from typing import Union
+from secrets import token_urlsafe
 from urllib.parse import urlparse, parse_qs
+
+from .token import Token
+from .expiring import Credentials
 from .refreshing import RefreshingToken, RefreshingCredentials
+
+
+def gen_state(n_bytes: int = 32) -> str:
+    """
+    Generate state to use in user authorisation.
+
+    The generated state is random and URL-safe.
+    It is generated using :func:`secrets.token_urlsafe`.
+    """
+    return token_urlsafe(n_bytes)
+
+
+def _parse_url_param(url: str, param: str) -> str:
+    query = urlparse(url).query
+    code = parse_qs(query).get(param, None)
+
+    if code is None:
+        raise KeyError(f'Parameter `{param}` not available!')
+    elif len(code) > 1:
+        raise KeyError(f'Multiple values for `{param}`!')
+
+    return code[0]
 
 
 def parse_code_from_url(url: str) -> str:
@@ -18,15 +45,119 @@ def parse_code_from_url(url: str) -> str:
     KeyError
         if 'code' is not available or has multiple values
     """
-    query = urlparse(url).query
-    code = parse_qs(query).get('code', None)
+    return _parse_url_param(url, 'code')
 
-    if code is None:
-        raise KeyError('Parameter `code` not available!')
-    elif len(code) > 1:
-        raise KeyError('Multiple values for `code`!')
 
-    return code[0]
+def parse_state_from_url(url: str) -> str:
+    """
+    Parse an URL for parameter 'state'.
+
+    Returns
+    -------
+    str
+        value of 'state'
+
+    Raises
+    ------
+    KeyError
+        if 'state' is not available or has multiple values
+    """
+    return _parse_url_param(url, 'state')
+
+
+class UserAuth:
+    """
+    Implement user authorisation flow.
+
+    Implements all steps and security checks for user authorisation.
+    The responsibility of the caller is to redirect a user to the given URL
+    and provide the resulting redirect URI or its parameters.
+
+    Parameters
+    ----------
+    cred
+        credentials client
+    scope
+        token privileges, accepts a :class:`Scope`, a single :class:`scope`,
+        a list of :class:`scopes <scope>` and strings for :class:`Scope`,
+        or a space-separated list of scopes as a string
+
+    Attributes
+    ----------
+    url: str
+        address to redirect a user to for authorisation
+    state: str
+        generated additional state
+
+    Examples
+    --------
+    .. code:: python
+
+        auth = tk.UserAuth(cred, scope)
+
+        # Redirect user to auth.url and parse parameters
+        code, state = ...
+        token = auth.request_token(code, state)
+
+        # Or leave parsing to UserAuth
+        redirected = ...
+        token = auth.request_token(url=redirected)
+    """
+
+    def __init__(self, cred: Union[Credentials, RefreshingCredentials], scope=None):
+        self._cred = cred
+        self.state = gen_state()
+        self.url = self._cred.user_authorisation_url(
+            scope, self.state, show_dialog=True
+        )
+
+    def __repr__(self):
+        options = [
+            f'cred={self._cred!r}',
+            f'url={self.url!r}',
+            f'state={self.state!r}',
+        ]
+        return type(self).__name__ + '(' + ', '.join(options) + ')'
+
+    def request_token(
+        self,
+        code: str = None,
+        state: str = None,
+        url: str = None,
+    ) -> Union[Token, RefreshingToken]:
+        """
+        Verify state consistency and request token.
+
+        Parameters
+        ----------
+        code
+            code from redirect parameters, required if url was not specified
+        state
+            state from redirect parameters, required if url was not specified
+        url
+            if specified, code and state are parsed from this URL instead
+
+        Returns
+        -------
+        Union[Token, RefreshingToken]
+            access token
+
+        Raises
+        ------
+        AssertionError
+            if state is inconsistent
+        """
+        if url is not None:
+            code = parse_code_from_url(url)
+            state = parse_state_from_url(url)
+
+        if self.state != state:
+            raise AssertionError(
+                'Inconsistent state!'
+                f' Expected `{self.state}`, got `{state}`.'
+            )
+
+        return self._cred.request_user_token(code)
 
 
 def request_client_token(
@@ -59,10 +190,10 @@ def prompt_for_user_token(
         scope=None
 ) -> RefreshingToken:
     """
-    Prompt for manual authentication.
+    Prompt for manual authorisation.
 
     Open a web browser for the user to log in with Spotify.
-    Prompt to paste the URL after logging in to parse the `code` URL parameter.
+    Prompt to paste the URL after logging in to complete authorisation.
 
     Parameters
     ----------
@@ -81,15 +212,19 @@ def prompt_for_user_token(
     -------
     RefreshingToken
         automatically refreshing user token
+
+    Raises
+    ------
+    AssertionError
+        if state is inconsistent
     """
     cred = RefreshingCredentials(client_id, client_secret, redirect_uri)
-    url = cred.user_authorisation_url(scope, show_dialog=True)
+    auth = UserAuth(cred, scope=scope)
 
     print('Opening browser for Spotify login...')
-    webbrowser.open(url)
+    webbrowser.open(auth.url)
     redirected = input('Please paste redirect URL: ').strip()
-    code = parse_code_from_url(redirected)
-    return cred.request_user_token(code)
+    return auth.request_token(url=redirected)
 
 
 def refresh_user_token(