+++ tests/auth/util.py
@@ -2,8 +2,10 @@ import pytest
 from unittest.mock import MagicMock, patch
 
 from tekore import (
+    UserAuth,
     RefreshingToken,
     parse_code_from_url,
+    parse_state_from_url,
     prompt_for_user_token,
     refresh_user_token,
     request_client_token,
@@ -28,20 +30,66 @@ class TestParseCodeFromURL:
         assert r == '1'
 
 
+class TestParseStateFromURL:
+    def test_empty_url_raises(self):
+        with pytest.raises(KeyError):
+            parse_state_from_url('')
+
+    def test_no_code_raises(self):
+        with pytest.raises(KeyError):
+            parse_state_from_url('http://example.com')
+
+    def test_multiple_codes_raises(self):
+        with pytest.raises(KeyError):
+            parse_state_from_url('http://example.com?state=1&state=2')
+
+    def test_single_code_returned(self):
+        r = parse_state_from_url('http://example.com?state=1')
+        assert r == '1'
+
+
+def make_user_auth():
+    cred = MagicMock()
+    cred.user_authorisation_url.return_value = 'http://url.com'
+    cred.request_user_token.return_value = 'token'
+    return UserAuth(cred, 'scope')
+
+
+class TestUserAuth:
+    def test_repr(self):
+        auth = make_user_auth()
+        assert repr(auth).startswith('UserAuth(')
+
+    def test_user_auth_correct_state(self):
+        auth = make_user_auth()
+        assert auth.request_token('code', auth.state)
+
+    def test_user_auth_parse_from_url(self):
+        auth = make_user_auth()
+        url = f'http://redirect.com?code=code&state={auth.state}'
+        assert auth.request_token(url=url) == 'token'
+
+    def test_user_auth_incorrect_state_raises(self):
+        auth = make_user_auth()
+        with pytest.raises(AssertionError):
+            auth.request_token('code', 'wrong-state')
+
+
 class TestTokenUtilityFunctions:
     def test_prompt_for_user_token(self):
         cred = MagicMock()
         cred.authorisation_url.return_value = 'http://example.com'
-        cred.request_access_token.return_value = MagicMock()
-        input_ = MagicMock(return_value='http://example.com?code=1')
+        input_ = MagicMock(return_value='http://example.com?code=1&state=s')
+        state = MagicMock(return_value='s')
+        util_mod = 'tekore._auth.util'
         with patch('tekore._auth.refreshing.Credentials', cred),\
-                patch('tekore._auth.util.webbrowser', MagicMock()),\
-                patch('tekore._auth.util.input', input_),\
-                patch('tekore._auth.util.print', MagicMock()):
-            token = prompt_for_user_token('', '', '')
+                patch(util_mod + '.webbrowser', MagicMock()),\
+                patch(util_mod + '.input', input_),\
+                patch(util_mod + '.print', MagicMock()),\
+                patch(util_mod + '.gen_state', state):
+            prompt_for_user_token('', '', '')
 
         input_.assert_called_once()
-        assert isinstance(token, RefreshingToken)
 
     def test_request_refreshed_token_returns_refreshing_token(
             self, app_env, user_refresh